backend/requirements.txt
# Django and extensions
Django==4.2.0
django-cors-headers==4.3.1
django-stubs==4.2.7
djangorestframework==3.14.0
# GraphQL
graphene-django==3.0.0
django-graphql-jwt==0.4.0
django-filter==23.5

# Cache/Redis
redis==5.0.1
django-redis==5.4.0

# Email
django-anymail==10.2 

# Image Processing
Pillow==10.2.0  # For ImageField

# Development tools
pylint-django==2.5.5
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9

# Testing
pytest==7.4.4
pytest-django==4.7.0

# Event Broker
confluent-kafka==2.1.1
